requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.2
fastapi-cli==0.0.5
fastjsonschema==2.20.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
yarl==1.15.1
